requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
openai>=1.30.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=2.0.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
pymongo>=4.6.0
tiktoken>=0.7.0
